Move tests to json_path
