+++ lib/segment/src/common/utils.rs
@@ -115,721 +115,6 @@ impl<T: JsonSchema> JsonSchema for MaybeOneOrMany<T> {
 }
 
 #[cfg(test)]
-mod jsonpath_tests {
-    use super::*;
-    use crate::json_path::JsonPath;
-
-    fn json(str: &str) -> serde_json::Map<String, Value> {
-        serde_json::from_str(str).unwrap()
-    }
-
-    #[test]
-    fn test_get_nested_value_from_json_map() {
-        let map = serde_json::from_str::<serde_json::Map<String, Value>>(
-            r#"
-            {
-                "a": {
-                    "b": {
-                        "c": 1
-                    }
-                },
-                "d": 2
-            }
-            "#,
-        )
-        .unwrap();
-
-        assert_eq!(
-            JsonPath::new("a.b").value_get(&map).into_vec(),
-            vec![&Value::Object(serde_json::Map::from_iter(vec![(
-                "c".to_string(),
-                Value::Number(1.into())
-            )]))]
-        );
-
-        // going deeper
-        assert_eq!(
-            JsonPath::new("a.b.c").value_get(&map).into_vec(),
-            vec![&Value::Number(1.into())]
-        );
-
-        // missing path
-        assert!(check_is_empty(
-            JsonPath::new("a.b.c.d").value_get(&map).iter().copied()
-        ));
-    }
-
-    #[test]
-    fn test_is_empty() {
-        let map = serde_json::from_str::<serde_json::Map<String, Value>>(
-            r#"
-                {
-                   "a": [
-                     { "b": 1 },
-                     { "b": 2 },
-                     { "b": null },
-                     { "d": [] },
-                     { "d": [] },
-                     { "f": null }
-                   ]
-                }
-            "#,
-        )
-        .unwrap();
-        let multivalue = JsonPath::new("a[].b").value_get(&map);
-        let is_empty = check_is_empty(multivalue.iter().copied());
-
-        assert!(!is_empty, "a[].b is not empty");
-
-        let multivalue = JsonPath::new("a[].c").value_get(&map);
-        let is_empty = check_is_empty(multivalue.iter().copied());
-
-        assert!(is_empty, "a[].c is empty");
-
-        let multivalue = JsonPath::new("a[].d").value_get(&map);
-        let is_empty = check_is_empty(multivalue.iter().copied());
-        assert!(is_empty, "a[].d is empty");
-
-        let multivalue = JsonPath::new("a[].f").value_get(&map);
-        let is_empty = check_is_empty(multivalue.iter().copied());
-        assert!(is_empty, "a[].f is empty");
-    }
-
-    #[test]
-    fn test_get_nested_array_value_from_json_map() {
-        let map = serde_json::from_str::<serde_json::Map<String, Value>>(
-            r#"
-            {
-                "a": {
-                    "b": [
-                        { "c": 1 },
-                        { "c": 2 },
-                        { "d": { "e": 3 } }
-                    ]
-                },
-                "f": 3,
-                "g": ["g0", "g1", "g2"]
-            }
-            "#,
-        )
-        .unwrap();
-
-        // get JSON array
-        assert_eq!(
-            JsonPath::new("a.b").value_get(&map).into_vec(),
-            vec![&Value::Array(vec![
-                Value::Object(serde_json::Map::from_iter(vec![(
-                    "c".to_string(),
-                    Value::Number(1.into())
-                )])),
-                Value::Object(serde_json::Map::from_iter(vec![(
-                    "c".to_string(),
-                    Value::Number(2.into())
-                )])),
-                Value::Object(serde_json::Map::from_iter(vec![(
-                    "d".to_string(),
-                    Value::Object(serde_json::Map::from_iter(vec![(
-                        "e".to_string(),
-                        Value::Number(3.into())
-                    )]))
-                )]))
-            ])]
-        );
-
-        // a.b[] extract all elements from array
-        assert_eq!(
-            JsonPath::new("a.b[]").value_get(&map).into_vec(),
-            vec![
-                &Value::Object(serde_json::Map::from_iter(vec![(
-                    "c".to_string(),
-                    Value::Number(1.into())
-                )])),
-                &Value::Object(serde_json::Map::from_iter(vec![(
-                    "c".to_string(),
-                    Value::Number(2.into())
-                )])),
-                &Value::Object(serde_json::Map::from_iter(vec![(
-                    "d".to_string(),
-                    Value::Object(serde_json::Map::from_iter(vec![(
-                        "e".to_string(),
-                        Value::Number(3.into())
-                    )]))
-                )]))
-            ]
-        );
-
-        // project scalar field through array
-        assert_eq!(
-            JsonPath::new("a.b[].c").value_get(&map).into_vec(),
-            vec![&Value::Number(1.into()), &Value::Number(2.into())]
-        );
-
-        // project object field through array
-        assert_eq!(
-            JsonPath::new("a.b[].d").value_get(&map).into_vec(),
-            vec![&Value::Object(serde_json::Map::from_iter(vec![(
-                "e".to_string(),
-                Value::Number(3.into())
-            )]))]
-        );
-
-        // select scalar element from array
-        assert_eq!(
-            JsonPath::new("a.b[0]").value_get(&map).into_vec(),
-            vec![&Value::Object(serde_json::Map::from_iter(vec![(
-                "c".to_string(),
-                Value::Number(1.into())
-            )]))]
-        );
-
-        // select scalar object from array different index
-        assert_eq!(
-            JsonPath::new("a.b[1]").value_get(&map).into_vec(),
-            vec![&Value::Object(serde_json::Map::from_iter(vec![(
-                "c".to_string(),
-                Value::Number(2.into())
-            )]))]
-        );
-
-        // select field element from array different index
-        assert_eq!(
-            JsonPath::new("a.b[1].c").value_get(&map).into_vec(),
-            vec![&Value::Number(2.into())]
-        );
-
-        // select scalar element from array different index
-        assert_eq!(
-            JsonPath::new("g[2]").value_get(&map).into_vec(),
-            vec![&Value::String("g2".to_string())]
-        );
-
-        // select object element from array
-        assert_eq!(
-            JsonPath::new("a.b[2]").value_get(&map).into_vec(),
-            vec![&Value::Object(serde_json::Map::from_iter(vec![(
-                "d".to_string(),
-                Value::Object(serde_json::Map::from_iter(vec![(
-                    "e".to_string(),
-                    Value::Number(3.into())
-                )]))
-            )]))]
-        );
-
-        // select out of bound index from array
-        assert!(check_is_empty(
-            JsonPath::new("a.b[3]").value_get(&map).iter().copied()
-        ));
-    }
-
-    #[test]
-    fn test_get_deeply_nested_array_value_from_json_map() {
-        let map = serde_json::from_str::<serde_json::Map<String, Value>>(
-            r#"
-            {
-                "arr1": [
-                    {
-                        "arr2": [
-                            {"a": 1, "b": 2}
-                        ]
-                    },
-                    {
-                        "arr2": [
-                            {"a": 3, "b": 4},
-                            {"a": 5, "b": 6}
-                        ]
-                    }
-                ]
-            }
-            "#,
-        )
-        .unwrap();
-
-        // extract and flatten all elements from arrays
-        assert_eq!(
-            JsonPath::new("arr1[].arr2[].a").value_get(&map).into_vec(),
-            vec![
-                &Value::Number(1.into()),
-                &Value::Number(3.into()),
-                &Value::Number(5.into()),
-            ]
-        );
-    }
-
-    #[test]
-    fn test_no_flatten_array_value_from_json_map() {
-        let map = serde_json::from_str::<serde_json::Map<String, Value>>(
-            r#"
-            {
-                "arr": [
-                    { "a": [1, 2, 3] },
-                    { "a": 4 },
-                    { "b": 5 }
-                ]
-            }
-            "#,
-        )
-        .unwrap();
-
-        // extract and retain structure for arrays arrays
-        assert_eq!(
-            JsonPath::new("arr[].a").value_get(&map).into_vec(),
-            vec![
-                &Value::Array(vec![
-                    Value::Number(1.into()),
-                    Value::Number(2.into()),
-                    Value::Number(3.into()),
-                ]),
-                &Value::Number(4.into()),
-            ]
-        );
-
-        // expect an array as leaf, ignore non arrays
-        assert_eq!(
-            JsonPath::new("arr[].a[]").value_get(&map).into_vec(),
-            vec![
-                &Value::Number(1.into()),
-                &Value::Number(2.into()),
-                &Value::Number(3.into()),
-            ]
-        );
-    }
-
-    #[test]
-    fn test_get_null_and_absent_values() {
-        let map = serde_json::from_str::<serde_json::Map<String, Value>>(
-            r#"
-            {
-                "a": null,
-                "b": [null, null],
-                "c": []
-            }
-            "#,
-        )
-        .unwrap();
-
-        assert_eq!(
-            JsonPath::new("a").value_get(&map).as_slice(),
-            &[&Value::Null],
-        );
-
-        assert!(JsonPath::new("a[]").value_get(&map).is_empty());
-
-        assert_eq!(
-            JsonPath::new("b").value_get(&map).as_slice(),
-            &[&Value::Array(vec![Value::Null, Value::Null])],
-        );
-
-        assert_eq!(
-            JsonPath::new("b[]").value_get(&map).as_slice(),
-            &[&Value::Null, &Value::Null],
-        );
-
-        assert_eq!(
-            JsonPath::new("c").value_get(&map).as_slice(),
-            &[&Value::Array(vec![])],
-        );
-
-        assert!(JsonPath::new("c[]").value_get(&map).is_empty());
-
-        assert!(JsonPath::new("d").value_get(&map).is_empty());
-
-        assert!(JsonPath::new("d[]").value_get(&map).is_empty());
-    }
-
-    #[test]
-    fn test_filter_json() {
-        let map = serde_json::from_str::<serde_json::Map<String, Value>>(
-            r#"
-            {
-                "a": {
-                    "b": [
-                        { "c": 1 },
-                        { "c": 2 },
-                        { "d": { "e": 3 } }
-                    ]
-                },
-                "f": 3,
-                "g": ["g0", "g1", "g2"]
-            }
-            "#,
-        )
-        .unwrap();
-
-        let res = JsonPath::value_filter(&map, |path, _value| {
-            let path = path.to_string();
-            path.starts_with("a.b[].c") || "a.b[].c".starts_with(&path)
-        });
-
-        assert_eq!(
-            res,
-            serde_json::from_str::<serde_json::Map<String, Value>>(
-                r#"
-                {
-                    "a": {
-                        "b": [
-                            { "c": 1 },
-                            { "c": 2 },
-                            {}
-                        ]
-                    }
-                }
-                "#,
-            )
-            .unwrap()
-        );
-    }
-
-    #[test]
-    fn test_check_include_pattern() {
-        assert!(JsonPath::new("a.b.c").check_include_pattern(&JsonPath::new("a.b.c")));
-        assert!(JsonPath::new("a.b.c").check_include_pattern(&JsonPath::new("a.b")));
-        assert!(!JsonPath::new("a.b.c").check_include_pattern(&JsonPath::new("a.b.d")));
-        assert!(JsonPath::new("a.b.c").check_include_pattern(&JsonPath::new("a")));
-        assert!(JsonPath::new("a").check_include_pattern(&JsonPath::new("a.d")));
-    }
-
-    #[test]
-    fn test_check_exclude_pattern() {
-        assert!(JsonPath::new("a.b.c").check_exclude_pattern(&JsonPath::new("a.b.c")));
-        assert!(!JsonPath::new("a.b.c").check_exclude_pattern(&JsonPath::new("a.b")));
-        assert!(!JsonPath::new("a.b.c").check_exclude_pattern(&JsonPath::new("a.b.d")));
-        assert!(!JsonPath::new("a.b.c").check_exclude_pattern(&JsonPath::new("a")));
-        assert!(JsonPath::new("a").check_exclude_pattern(&JsonPath::new("a.d")));
-    }
-
-    #[test]
-    fn test_set_value_to_json_with_empty_key() {
-        let mut map = serde_json::from_str::<serde_json::Map<String, Value>>(
-            r#"
-            {
-                "a": {
-                    "b": [
-                        { "c": 1 },
-                        { "c": 2 },
-                        { "d": { "e": 3 } }
-                    ]
-                },
-                "f": 3,
-                "g": ["g0", "g1", "g2"]
-            }
-            "#,
-        )
-        .unwrap();
-
-        let src = serde_json::from_str::<serde_json::Map<String, Value>>(
-            r#"
-            { "c": 5 }
-            "#,
-        )
-        .unwrap();
-
-        JsonPath::value_set(None, &mut map, &src);
-
-        assert_eq!(
-            map,
-            serde_json::from_str::<serde_json::Map<String, Value>>(
-                r#"
-                {
-                    "a": {
-                        "b": [
-                            { "c": 1 },
-                            { "c": 2 },
-                            { "d": { "e": 3 } }
-                        ]
-                    },
-                    "f": 3,
-                    "g": ["g0", "g1", "g2"],
-                    "c": 5
-                }
-                "#,
-            )
-            .unwrap()
-        );
-    }
-
-    #[test]
-    fn test_set_value_to_json_with_one_level_key() {
-        let mut map = serde_json::from_str::<serde_json::Map<String, Value>>(
-            r#"
-            {
-                "a": {
-                    "b": [
-                        { "c": 1 },
-                        { "c": 2 },
-                        { "d": { "e": 3 } }
-                    ]
-                },
-                "f": 3,
-                "g": ["g0", "g1", "g2"]
-            }
-            "#,
-        )
-        .unwrap();
-
-        let src = serde_json::from_str::<serde_json::Map<String, Value>>(
-            r#"
-            { "b": 5 }
-            "#,
-        )
-        .unwrap();
-
-        JsonPath::value_set(Some(&JsonPath::new("a")), &mut map, &src);
-
-        assert_eq!(
-            map,
-            serde_json::from_str::<serde_json::Map<String, Value>>(
-                r#"
-                {
-                    "a": {
-                        "b": 5
-                    },
-                    "f": 3,
-                    "g": ["g0", "g1", "g2"]
-                }
-                "#,
-            )
-            .unwrap()
-        );
-    }
-
-    #[test]
-    fn test_set_value_to_json_with_array_index() {
-        let mut map = serde_json::from_str::<serde_json::Map<String, Value>>(
-            r#"
-            {
-                "a": {
-                    "b": [
-                        { "c": 1 },
-                        { "c": 2 },
-                        { "d": { "e": 3 } }
-                    ]
-                },
-                "f": 3,
-                "g": ["g0", "g1", "g2"]
-            }
-            "#,
-        )
-        .unwrap();
-
-        let src = serde_json::from_str::<serde_json::Map<String, Value>>(
-            r#"
-            { "c": 5 }
-            "#,
-        )
-        .unwrap();
-
-        JsonPath::value_set(Some(&JsonPath::new("a.b[1]")), &mut map, &src);
-
-        assert_eq!(
-            map,
-            serde_json::from_str::<serde_json::Map<String, Value>>(
-                r#"
-                {
-                    "a": {
-                        "b": [
-                            { "c": 1 },
-                            { "c": 5 },
-                            { "d": { "e": 3 } }
-                        ]
-                    },
-                    "f": 3,
-                    "g": ["g0", "g1", "g2"]
-                }
-                "#,
-            )
-            .unwrap()
-        );
-    }
-
-    #[test]
-    fn test_set_value_to_json_with_empty_src() {
-        let mut map = serde_json::from_str::<serde_json::Map<String, Value>>(
-            r#"
-            {
-                "a": {
-                    "b": [
-                        { "c": 1 },
-                        { "c": 2 },
-                        { "d": { "e": 3 } }
-                    ]
-                },
-                "f": 3,
-                "g": ["g0", "g1", "g2"]
-            }
-            "#,
-        )
-        .unwrap();
-
-        let src = serde_json::from_str::<serde_json::Map<String, Value>>(
-            r#"
-            {}
-            "#,
-        )
-        .unwrap();
-
-        JsonPath::value_set(Some(&JsonPath::new("a.b[1]")), &mut map, &src);
-
-        assert_eq!(
-            map,
-            serde_json::from_str::<serde_json::Map<String, Value>>(
-                r#"
-                {
-                    "a": {
-                        "b": [
-                            { "c": 1 },
-                            { "c": 2 },
-                            { "d": { "e": 3 } }
-                        ]
-                    },
-                    "f": 3,
-                    "g": ["g0", "g1", "g2"]
-                }
-                "#,
-            )
-            .unwrap()
-        );
-    }
-
-    #[test]
-    fn test_set_value_to_json_with_empty_dest() {
-        let mut map = serde_json::from_str::<serde_json::Map<String, Value>>(
-            r#"
-            {
-            }
-            "#,
-        )
-        .unwrap();
-
-        let src = serde_json::from_str::<serde_json::Map<String, Value>>(
-            r#"
-            {"c": 1}
-            "#,
-        )
-        .unwrap();
-
-        JsonPath::value_set(None, &mut map, &src);
-
-        assert_eq!(
-            map,
-            serde_json::from_str::<serde_json::Map<String, Value>>(
-                r#"
-                {
-                    "c": 1
-                }
-                "#,
-            )
-            .unwrap()
-        );
-    }
-
-    #[test]
-    fn test_set_value_to_json_with_empty_dest_nested_key() {
-        let mut map = serde_json::from_str::<serde_json::Map<String, Value>>(
-            r#"
-            {
-            }
-            "#,
-        )
-        .unwrap();
-
-        let src = serde_json::from_str::<serde_json::Map<String, Value>>(
-            r#"
-            {"c": 1}
-            "#,
-        )
-        .unwrap();
-
-        JsonPath::value_set(Some(&JsonPath::new("key1.key2")), &mut map, &src);
-
-        assert_eq!(
-            map,
-            serde_json::from_str::<serde_json::Map<String, Value>>(
-                r#"
-                {
-                    "key1": {"key2": { "c": 1 } }
-                }
-                "#,
-            )
-            .unwrap()
-        );
-    }
-
-    #[test]
-    fn test_set_value_to_json_with_empty_dest_nested_array_index_key() {
-        let mut map = json("{}");
-        let src = json(r#" {"c": 1} "#);
-        JsonPath::value_set(Some(&JsonPath::new("key1.key2[3]")), &mut map, &src);
-        assert_eq!(map, json(r#" {"key1": {"key2": []}} "#));
-
-        let mut map = json("{}");
-        let src = json(r#" {"c": 1} "#);
-        JsonPath::value_set(Some(&JsonPath::new("key1.key2[0]")), &mut map, &src);
-        assert_eq!(map, json(r#" {"key1": {"key2": []}} "#));
-    }
-
-    #[test]
-    fn test_expand_payload_with_non_existing_array() {
-        let mut map = serde_json::from_str::<serde_json::Map<String, Value>>(
-            r#"
-            {
-            }
-            "#,
-        )
-        .unwrap();
-
-        let src = serde_json::from_str::<serde_json::Map<String, Value>>(
-            r#"
-            {"c": 1}
-            "#,
-        )
-        .unwrap();
-
-        JsonPath::value_set(Some(&JsonPath::new("key1.key2[].key3")), &mut map, &src);
-
-        assert_eq!(
-            map,
-            serde_json::from_str::<serde_json::Map<String, Value>>(
-                r#"
-                {
-                    "key1": { "key2": [] }
-                }
-                "#,
-            )
-            .unwrap()
-        );
-    }
-
-    #[test]
-    fn test_replace_scalar_key_with_object() {
-        let mut map = serde_json::from_str::<serde_json::Map<String, Value>>(
-            r#"
-            {"a": 10}
-            "#,
-        )
-        .unwrap();
-
-        let src = serde_json::from_str::<serde_json::Map<String, Value>>(
-            r#"
-            {"x": 1}
-            "#,
-        )
-        .unwrap();
-
-        JsonPath::value_set(Some(&JsonPath::new("a.b.c")), &mut map, &src);
-
-        assert_eq!(
-            map,
-            serde_json::from_str::<serde_json::Map<String, Value>>(
-                r#"
-                {"a": {"b": {"c": {"x": 1}}}}
-                "#,
-            )
-            .unwrap()
-        );
-    }
-}
-
-#[cfg(test)]
 mod tests {
     use schemars::{schema_for, JsonSchema};
     use serde::{Deserialize, Serialize};
